// src/common/enums/events.enum.ts
export enum InAppEvents {
  SEND_MESSAGE = 'SEND_MESSAGE',
  MESSAGE_STATUS_UPDATED = 'MESSAGE_STATUS_UPDATED',
  BOT_REPLY_REQUIRED = 'BOT_REPLY_REQUIRED',
  BOT_STATUS_CHANGED = 'BOT_STATUS_CHANGED',
  BOT_STREAM_DELTA = 'BOT_STREAM_DELTA',
  JOIN_ROOM = 'JOIN_ROOM',
}

export enum WebSocketEvents {
  SEND_MESSAGE = 'SEND_MESSAGE',
  MESSAGE_STATUS_UPDATED = 'MESSAGE_STATUS_UPDATED',
  BOT_STATUS_CHANGED = 'BOT_STATUS_CHANGED',
  BOT_STREAM_DELTA = 'BOT_STREAM_DELTA',
  WIDGET_ERROR = 'WIDGET_ERROR',
  JOIN_ROOM = 'JOIN_ROOM',
}

export enum BotReplyStatus {
  THINKING = 'THINKING',
  DONE = 'DONE',
  ERROR = 'ERROR',
}